add triggers to allow for hooking into different steps of the process
In some instances, we want to add some conditionals to modify areas
like metadata. So adding hooks allow us to add points where we can
hook in and modify the state for future steps.

One of these instances are with some plans, where we may want to
only process a single interface. So we a hook after initialization
can be triggered to remove any additional interfaces that may have
been provided from the metadata service.