add support for appending to files
Some customers with customer images have some custom files. One file
in particular we've seen is `/etc/modules`. So instead of overwriting
the file, we would like to append to the file.

This change allows us to add a `file_mode` option to tasks, modifying
the `mode` attribute in `open()`. By default `w` is used, however it
can be set to `a` to append to the file.