resolve symlink paths before writing file
It's possible that an os could have a path already exist as a symlink without the linked
file existing. In python this results in an error `FileNotFoundError: [Errno 2] No such file or directory`

To get around this, we'll resolve any symlink paths first, before attempting to write to the file.