include all interfaces in mactab
If more interfaces exist on the host than is registered in metadata,
and if the interfaces in metadata are not ordered first on the host.

Example:
  metadata:
    eth0: macA
    eth1: macB
  host:
    eth0: macC
    eth1: macD
    eth2: macA
    eth3: macB

The `nameif` tool which reads the mactab, will attempt to rename
eth2 to eth0 and eth3 to eth1 however eth0 and eth1 are already
taken by interfaces it doesn't know about and therefore cannot
apply the name fix.

This change resolves that by setting the interfaces defined in
metadata first and then defines all remaining interfaces as well
to ensure the metadata defined interfaces can be properly set.

Signed-off-by: Mike Mason <mason@packet.com>