Clean up final appearance of /etc/network/interfaces files
This is a minor refactor that I think makes things much easier to grok and
thus worth it. dns-nameserver goes up with the ip config stuff instead after
bond config. All the bond config stuff a grouped together w/o spaces. I also
tweaked the templates so that new lines are part of {% if %} blocks so we
don't have to deal with arbritary empty lines in the tests. This makes the
source templates a little busier because they don't have the empty lines
between the {% if %} blocks but I think its worth it since it makes writing
tests easier.