Avoid duplicating expected results for debian/ubuntu tests
I figured out what needed to be done to share as much of the expected
results as possible. This is going to come very handy when we get
variable number of nics which requires the results to built up in pieces
(to avoid using jinja template to verify jinja template...).