Ensure bond0 has only the requested nics as members for debianish oses
This has been a long time bug that is very hard to reproduce due to
depending on what order the nics come up on the network. Quoting Tim
Jackson:

> If all 4 nics are in the same bond, the remote (partner lacp id) has
> to match for all 4, so if say eth1 [recte eth2] comes up first on the
> server side, anything in the other bond from the ToR side will fail to
> actually bond due to a mismatch in the partner key.

This is only happening in Debian because Ubuntu has the `pre-up sleep 4`
for all the other nics so eth0 comes up first which causes the "main"
bond to be correct and thus networking works well enough to phone-home
and activate.