distros: Add bonds to jinja2 context
Will be used immediately to clean up the debian template a little bit.
Necessary for a future commit that will handle multiple bonds. This is
only coming in now because redhatish distros have one file per bond and
thus we iterate over the bonds in python instead of in jinja.