fix: AlpineBonded ENI template bad bond options
When both bond-master and bond-slaves are present, bond-slaves is
ignored. This leads to the bond-* options associated with the
bond-slaves stanza to be ignored as well. We observed that this led to
provisioned Alpine linux machines to have balance-rr (mode 3) bonding mode
instead of 802.3ad (mode 4) bonding mode. This caused packets to be sent
out-of-order, leading to TCP/IP congestion control due to packet loss.

Removing the bond-master option from the ethX stanzas fixes this issue
and ensures the bonding driver uses our declared bond-* options in the
bond stanzas.